Processes 100% funtional

+++ resources/views/registrationProcess/pwcnm_registration_process/index.blade.php
@@ -4,20 +4,20 @@
     <div class="container">
         <div class="row">
 
-            <div class="col-md-9">
+            <div class="col-md-9 col-md-offset-1">
                 <div class="panel panel-default">
-                    <div class="panel-heading">Pwcnm_registration_process</div>
+                    <div class="panel-heading">@lang('form.processes')</div>
                     <div class="panel-body">
-                        <a href="{{ url('/registrationProcess/pwcnm_registration_process/create') }}" class="btn btn-success btn-sm" title="Add New pwcnm_registration_process">
-                            <i class="fa fa-plus" aria-hidden="true"></i> Add New
+                        <a href="{{ url('/admin/procesos/create') }}" class="btn btn-success btn-sm" title=@lang('form.add')>
+                            <i class="fa fa-plus" aria-hidden="true"></i> @lang('form.add')
                         </a>
 
-                        {!! Form::open(['method' => 'GET', 'url' => '/registrationProcess/pwcnm_registration_process', 'class' => 'navbar-form navbar-right', 'role' => 'search'])  !!}
+                        {!! Form::open(['method' => 'GET', 'url' => '/admin/procesos', 'class' => 'navbar-form navbar-right', 'role' => 'search'])  !!}
                         <div class="input-group">
-                            <input type="text" class="form-control" name="search" placeholder="Search...">
+                            <input type="text" class="form-control" name="search" placeholder=@lang('form.search')>
                             <span class="input-group-btn">
                                 <button class="btn btn-default" type="submit">
-                                    <i class="fa fa-search"></i>
+                                    <span class="glyphicon glyphicon-search" aria-hidden="true"></span>
                                 </button>
                             </span>
                         </div>
@@ -29,29 +29,52 @@
                             <table class="table table-borderless">
                                 <thead>
                                     <tr>
-                                        <th>ID</th><th>Period</th><th>Anno</th><th>InitialDate</th><th>Actions</th>
+                                        <th>@lang('form.year')</th>
+                                        <th>@lang('form.period')</th>
+                                        <th>@lang('form.iniDate')</th>
+                                        <th>@lang('form.remDays')</th>
+                                        <th>@lang('form.act')</th>
                                     </tr>
                                 </thead>
                                 <tbody>
                                 @foreach($pwcnm_registration_process as $item)
                                     <tr>
-                                        <td>{{ $item->id }}</td>
-                                        <td>{{ $item->period }}</td><td>{{ $item->anno }}</td><td>{{ $item->InitialDate }}</td>
+                                        <td>{{ $item->anno }}</td>
+                                        <td>{{ $item->period }}</td>
+                                        <td>{{ $item->InitialDate }}</td>
                                         <td>
-                                            <a href="{{ url('/registrationProcess/pwcnm_registration_process/' . $item->id) }}" title="View pwcnm_registration_process"><button class="btn btn-info btn-xs"><i class="fa fa-eye" aria-hidden="true"></i> View</button></a>
-                                            <a href="{{ url('/registrationProcess/pwcnm_registration_process/' . $item->id . '/edit') }}" title="Edit pwcnm_registration_process"><button class="btn btn-primary btn-xs"><i class="fa fa-pencil-square-o" aria-hidden="true"></i> Edit</button></a>
+                                        @if ($item->FinalDate <= 0)
+                                            @lang('form.ended');
+                                        @else
+                                            {{$item->FinalDate}}
+                                        @endif
+
+                                        </td>
+                                        <td>
+                                            <a href="{{ url('/admin/procesos/' . $item->id) }}" title=@lang('form.view')>
+                                                <button class="btn btn-info btn-xs">
+                                                    <span class="glyphicon glyphicon-eye-open" aria-hidden="true"></span>
+                                                </button>
+                                            </a>
+                                        @if ($item->FinalDate > 0)
+                                            <a href="{{ url('/admin/procesos/' . $item->id . '/edit') }}" title=@lang('form.edit')>
+                                                <button class="btn btn-primary btn-xs">
+                                                    <span class="glyphicon glyphicon-pencil" aria-hidden="true"></span>
+                                                </button>
+                                            </a>
                                             {!! Form::open([
                                                 'method'=>'DELETE',
-                                                'url' => ['/registrationProcess/pwcnm_registration_process', $item->id],
+                                                'url' => ['/admin/procesos', $item->id],
                                                 'style' => 'display:inline'
                                             ]) !!}
-                                                {!! Form::button('<i class="fa fa-trash-o" aria-hidden="true"></i> Delete', array(
+                                                {!! Form::button('<span class="glyphicon glyphicon-trash" aria-hidden="true"></span>', array(
                                                         'type' => 'submit',
                                                         'class' => 'btn btn-danger btn-xs',
-                                                        'title' => 'Delete pwcnm_registration_process',
-                                                        'onclick'=>'return confirm("Confirm delete?")'
+                                                        'title' => ''.trans('form.delete'),
+                                                        'onclick'=>'return confirm("'.@trans('form.comfirm').'")'
                                                 )) !!}
                                             {!! Form::close() !!}
+                                        @endif
                                         </td>
                                     </tr>
                                 @endforeach
